+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+-
+
+## [v0.6.0](https://github.com/lpc55/lpc55-hal/releases/tag/0.6.0) - 2026-06-23
+
 - Bump MSRV to 1.79
 - Remove `littlefs2_prince_filesystem!` macro because the implementation did not correctly determine the PRINCE regions to enable.
   Users should replace it with a custom implementation using the appropriate PRINCE regions for their configuration.